Move request performer to test module
